api 15 updates

+++ Midibard/Extensions/Dalamud/PartyExtensions.cs
@@ -8,15 +8,15 @@ namespace MidiBard.Extensions.Dalamud.Party;
 
 public static class PartyExtensions
 {
-    public static IPartyMember? GetMeAsPartyMember(this IPartyList partyList) => partyList.IsInParty() ? partyList.FirstOrDefault(i => i.ContentId == (long)DalamudApi.PlayerState.ContentId) : null;
+    public static IPartyMember? GetMeAsPartyMember(this IPartyList partyList) => partyList.IsInParty() ? partyList.FirstOrDefault(i => i.ContentId == DalamudApi.PlayerState.ContentId) : null;
     public static IPartyMember? GetPartyLeader(this IPartyList partyList) => partyList.IsInParty() ? partyList[(int)partyList.PartyLeaderIndex] : null;
     public static bool IsInParty(this IPartyList partyList) => partyList?.Length > 1;
     public static bool IsPartyLeader(this IPartyMember member) => DalamudApi.PartyList.IsInParty() && member != null && member.ContentId == DalamudApi.PartyList.GetPartyLeader()?.ContentId;
-    public static bool IsPartyLeader(this IPartyList partyList) => partyList.IsInParty() && (long)DalamudApi.PlayerState.ContentId == partyList.GetPartyLeader()?.ContentId;
-    public static IPartyMember? GetPartyMemberFromCid(this IPartyList partyList, long cid) => partyList.FirstOrDefault(i => i.ContentId == cid);
+    public static bool IsPartyLeader(this IPartyList partyList) => partyList.IsInParty() && DalamudApi.PlayerState.ContentId == partyList.GetPartyLeader()?.ContentId;
+    public static IPartyMember? GetPartyMemberFromCid(this IPartyList partyList, ulong cid) => partyList.FirstOrDefault(i => i.ContentId == cid);
     public static string NameAndWorld(this IPartyMember member) => $"{member?.Name}·{member?.World.ValueNullable?.Name.ToDalamudString().TextValue}";
 
-    public static (long Cid, string Name, string World) GetPartyMemberData(this IPartyMember member)
+    public static (ulong Cid, string Name, string World) GetPartyMemberData(this IPartyMember member)
     {
         var name = member?.Name.ToString() ?? "";
         var world = member?.World.ValueNullable?.Name.ToDalamudString().TextValue ?? "";
